+++ Gems/ScriptCanvas/Code/Builder/ScriptCanvasBuilderDataSystem.h
@@ -32,7 +32,10 @@ namespace ScriptCanvasBuilder
         AZ_TYPE_INFO(DataSystem, "{27B74209-319D-4A8C-B37D-F85EFA6D2FFA}");
         AZ_CLASS_ALLOCATOR(DataSystem, AZ::SystemAllocator, 0);
 
+        DataSystem();
+
     protected:
+        void CompileBuilderData(ScriptCanvasEditor::SourceHandle sourceHandle);
         void SourceFileChanged(AZStd::string relativePath, AZStd::string scanFolder, AZ::Uuid fileAssetId) override;
         void SourceFileRemoved(AZStd::string relativePath, AZStd::string scanFolder, AZ::Uuid fileAssetId) override;
         void SourceFileFailed(AZStd::string relativePath, AZStd::string scanFolder, AZ::Uuid fileAssetId) override;
